requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-import-export==4.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tablib==3.7.0
tzdata==2024.2
